feat(reroute): Add segmentSpeeds to reroute modifications
