+++ lib/components/modification/__tests__/reroute.js
@@ -19,7 +19,6 @@ describe('Component > Modification > Reroute', () => {
       },
       setMapState: jest.fn(),
       stops: [],
-      totalTravelTimeMinutes: 0,
       update: jest.fn()
     }
     const tree = renderer.create(
