+++ lib/components/modification/add-trip-pattern.js
@@ -1,9 +1,7 @@
 /** display an add trip pattern modification */
 
 import React, {PropTypes} from 'react'
-import lineDistance from 'turf-line-distance'
 
-import {DEFAULT_SEGMENT_SPEED} from '../../constants/timetables'
 import {Button} from '../buttons'
 import DeepEqual from '../deep-equal'
 import Icon from '../icon'
@@ -26,44 +24,10 @@ export default class AddTripPattern extends DeepEqual {
 
   state = getStateFromProps(this.props)
 
-  componentDidMount () {
-    const {modification} = this.props
-    if (modification.segments.length > 0 &&
-      modification.timetables.length > 0 &&
-      modification.timetables.some((tt) => tt.segmentSpeeds.length !== modification.segments.length)) {
-      this._setTimetableSegmentLengthsTo(modification.segments.length)
-    }
-  }
-
   componentWillReceiveProps (newProps) {
-    if (newProps.modification.segments !== this.props.modification.segments) {
-      this._setTimetableSegmentLengthsTo(newProps.modification.segments.length)
-    }
-
     this.setState({
       ...this.state,
       ...getStateFromProps(newProps)
-    })
-  }
-
-  _setTimetableSegmentLengthsTo (newLength) {
-    const {modification, update} = this.props
-    update({
-      timetables: modification.timetables.map((t) => {
-        if (newLength <= t.segmentSpeeds.length) {
-          return {
-            ...t,
-            segmentSpeeds: t.segmentSpeeds.slice(0, newLength)
-          }
-        } else {
-          const segmentSpeeds = t.segmentSpeeds
-          while (segmentSpeeds.length < newLength) segmentSpeeds.push(segmentSpeeds[0] || DEFAULT_SEGMENT_SPEED)
-          return {
-            ...t,
-            segmentSpeeds
-          }
-        }
-      })
     })
   }
 
@@ -137,8 +101,8 @@ export default class AddTripPattern extends DeepEqual {
   }
 
   render () {
-    const {modification, mapState, numberOfStops, update} = this.props
-    const {extendFromEnd, followRoad, segmentDistances} = this.state
+    const {modification, mapState, numberOfStops, segmentDistances, update} = this.props
+    const {extendFromEnd, followRoad} = this.state
     // not using state here as otherwise checkbox would toggle on and off as you're editing the stop spacing
     const autoCreateStops = modification.segments.length > 0 ? modification.segments[0].spacing > 0 : true
     const isEditing = mapState.modificationId === modification.id
@@ -205,13 +169,11 @@ export default class AddTripPattern extends DeepEqual {
 
 function getStateFromProps ({
   mapState,
-  modification,
-  numberOfStops
+  modification
 }) {
   return {
     extendFromEnd: mapState && mapState.modificationId === modification.id ? mapState.extendFromEnd : true,
     followRoad: mapState.followRoad,
-    segmentDistances: numberOfStops > 1 ? modification.segments.map((s) => lineDistance(s.geometry, 'kilometers')) : [],
     spacing: modification.segments.length > 0 ? modification.segments[0].spacing : DEFAULT_SPACING
   }
 }
