+++ lib/components/modification/editor.js
@@ -1,4 +1,5 @@
 import React, {Component, PropTypes} from 'react'
+import lineDistance from 'turf-line-distance'
 
 import AddTripPattern from './add-trip-pattern'
 import AdjustDwellTime from './adjust-dwell-time'
@@ -32,6 +33,25 @@ export default class ModificationEditor extends Component {
     mapState: PropTypes.object.isRequired,
     setActiveTrips: PropTypes.func.isRequired,
     setMapState: PropTypes.func.isRequired
+  }
+
+  state = {
+    stops: getStops(this.props.modification.segments),
+    segmentDistances: this.props.numberOfStops > 1
+      ? this.props.modification.segments.map((s) => lineDistance(s.geometry, 'kilometers'))
+      : []
+  }
+
+  componentWillReceiveProps ({modification}) {
+    const {segments} = modification
+    if (segments !== this.props.modification.segments) {
+      const stops = getStops(segments)
+      this.setState({
+        ...this.state,
+        stops,
+        segmentDistances: stops.length > 1 ? segments.map((s) => lineDistance(s.geometry, 'kilometers')) : []
+      })
+    }
   }
 
   _onNameChange = (e) => {
@@ -76,6 +96,7 @@ export default class ModificationEditor extends Component {
   render () {
     const {allVariants, feeds, feedsById, modification, mapState, name, setActiveTrips, setMapState, type, variants} = this.props
     const props = {
+      ...this.state,
       feeds,
       feedsById,
       mapState,
@@ -87,13 +108,10 @@ export default class ModificationEditor extends Component {
     }
 
     if (modification.segments) {
-      props.stops = getStops(modification.segments)
       props.numberOfStops = props.stops.length
       props.lastStopDistanceFromStart = props.numberOfStops > 0
         ? props.stops.slice(-1)[0].distanceFromStart
         : 0
-
-      props.totalTravelTimeMinutes = getTravelTimeMinutes(props)
     }
 
     const ModificationType = getComponentForType(type)
@@ -135,24 +153,6 @@ export default class ModificationEditor extends Component {
       </div>
     )
   }
-}
-
-function getTravelTimeMinutes ({lastStopDistanceFromStart, modification, stops}) {
-  const nStops = stops.length
-
-  let travelTime = 0
-  if (nStops > 0) {
-    travelTime = lastStopDistanceFromStart / 1000 / modification.speed * 60
-    // add the dwell times
-    const dwellMinutes = modification.dwell / 60
-    travelTime += nStops * dwellMinutes
-
-    // don't include dwell time at first and last stop of reroute
-    if (!modification.fromStop) travelTime -= dwellMinutes
-    if (!modification.toStop) travelTime -= dwellMinutes
-  }
-
-  return travelTime
 }
 
 function getComponentForType (type) {
