+++ lib/components/modification/reroute.js
@@ -6,6 +6,7 @@ import {Button} from '../buttons'
 import Icon from '../icon'
 import {Group, Number, Checkbox} from '../input'
 import SelectFeedRouteAndPatterns from './select-feed-route-and-patterns'
+import SegmentSpeeds from './segment-speeds'
 import messages from '../../utils/messages'
 import {ADD_STOPS, SINGLE_STOP_SELECTION} from '../scenario-map/state'
 
@@ -13,12 +14,11 @@ export default class Reroute extends Component {
   static propTypes = {
     feeds: PropTypes.array.isRequired,
     feedsById: PropTypes.object.isRequired,
-    lastStopDistanceFromStart: PropTypes.number.isRequired,
     modification: PropTypes.object.isRequired,
     mapState: PropTypes.object.isRequired,
+    segmentDistances: PropTypes.array.isRequired,
     setMapState: PropTypes.func.isRequired,
     stops: PropTypes.array.isRequired,
-    totalTravelTimeMinutes: PropTypes.number.isRequired,
     update: PropTypes.func.isRequired
   }
 
@@ -82,18 +82,6 @@ export default class Reroute extends Component {
     })
   }
 
-  setSpeed = (e) => {
-    this.props.update({
-      speed: parseInt(e.target.value, 10)
-    })
-  }
-
-  setDwell = (e) => {
-    this.props.update({
-      dwell: parseInt(e.target.value, 10)
-    })
-  }
-
   /** toggle whether stops should be created automatically */
   onAutoCreateStopsChange = (e) => {
     const spacing = e.target.checked ? 400 : 0
@@ -118,28 +106,14 @@ export default class Reroute extends Component {
     }
   }
 
-  setSpeedFromTravelTime = (e) => {
-    const {lastStopDistanceFromStart, modification, stops, update} = this.props
-    let nStops = stops.length
-    // figure out appropriate speed given this travel time
-    const travelTimeMinutes = parseFloat(e.target.value)
-    const travelTimeIsValid = travelTimeMinutes !== null && !isNaN(travelTimeMinutes)
-    if (travelTimeIsValid && nStops !== 0) {
-      // don't include dwells at first and last stops
-      if (modification.fromStop) nStops--
-      if (modification.toStop) nStops--
-      const totalDwellSeconds = nStops * modification.dwell
-      const travelTimeSeconds = (travelTimeMinutes * 60) - totalDwellSeconds
-      // figure out speed
-      const speedMps = lastStopDistanceFromStart / travelTimeSeconds
-      const speedKph = speedMps * 3600 / 1000
-      update({speed: speedKph})
-    }
-  }
-
   render () {
-    const {feeds, modification, totalTravelTimeMinutes} = this.props
+    const {feeds, modification, segmentDistances, stops, update} = this.props
     const {followRoad, isEditingAlignment, selectedFeed, spacing} = this.state
+
+    let numberOfStops = stops.length
+    // don't include dwells at first and last stops
+    if (modification.fromStop) numberOfStops--
+    if (modification.toStop) numberOfStops--
 
     // not using state here as otherwise checkbox would toggle on and off as you're editing the stop spacing
     const autoCreateStops = modification.segments.length > 0 ? modification.segments[0].spacing > 0 : true
@@ -154,26 +128,6 @@ export default class Reroute extends Component {
           />
 
         {this.renderStops()}
-
-        <Number
-          label='Speed'
-          onChange={this.setSpeed}
-          units='km/h'
-          value={modification.speed}
-          />
-        <Number
-          label='Dwell time'
-          onChange={this.setDwell}
-          step={1}
-          units='seconds'
-          value={modification.dwell}
-        />
-        <Number
-          label='Travel time'
-          onChange={this.setSpeedFromTravelTime}
-          units='minutes'
-          value={totalTravelTimeMinutes}
-          />
 
         {isEditingAlignment &&
           <div>
@@ -212,6 +166,14 @@ export default class Reroute extends Component {
             onChange={this.onStopSpacingChange}
             />
         }
+
+        <SegmentSpeeds
+          dwellTime={modification.dwellTime}
+          numberOfStops={numberOfStops}
+          segmentDistances={segmentDistances}
+          segmentSpeeds={modification.segmentSpeeds}
+          update={update}
+          />
 
       </div>
     )
