+++ lib/components/modification/timetables.js
@@ -1,4 +1,3 @@
-import sum from 'lodash.sum'
 import React, {PropTypes} from 'react'
 
 import {DEFAULT_SEGMENT_SPEED} from '../../constants/timetables'
@@ -54,12 +53,6 @@ export default class Timetables extends DeepEqual {
             update={this._update.bind(this, i)}
             remove={this._remove.bind(this, i)}
             segmentDistances={segmentDistances}
-            travelTime={getTravelTimeMinutes({
-              dwellTime: tt.dwellTime,
-              numberOfStops,
-              segmentDistances,
-              segmentSpeeds: tt.segmentSpeeds
-            })}
             timetable={tt}
             />
         })}
@@ -72,25 +65,5 @@ export default class Timetables extends DeepEqual {
         </Button>
       </div>
     )
-  }
-}
-
-function getTravelTimeMinutes ({
-  dwellTime, // seconds
-  numberOfStops,
-  segmentDistances, // km
-  segmentSpeeds // kph
-}) {
-  if (numberOfStops > 0) {
-    const segmentTimes = segmentDistances.map(
-      (segmentDistance, i) => (segmentDistance / (segmentSpeeds[i] || DEFAULT_SEGMENT_SPEED)))
-    const totalMovingTime = sum(segmentTimes) * 60
-
-    // add the dwell times
-    const dwellMinutes = dwellTime / 60
-    const dwellingTime = numberOfStops * dwellMinutes
-    return totalMovingTime + dwellingTime
-  } else {
-    return 0
   }
 }
