+++ lib/utils/modification.js
@@ -7,7 +7,6 @@ import {ADD_TRIP_PATTERN, ADJUST_DWELL_TIME, ADJUST_SPEED, CONVERT_TO_FREQUENCY,
 const debug = dbg('scenario-editor:modification')
 
 const DEFAULT_ADD_STOPS_DWELL = 0
-const DEFAULT_ADD_STOPS_SPEED = 15
 const DEFAULT_ADJUST_DWELL_TIME_VALUE = 30
 const DEFAULT_ADJUST_SPEED_SCALE = 1
 
@@ -30,12 +29,12 @@ export function create ({
     case REROUTE:
       return {
         ...base,
-        dwell: DEFAULT_ADD_STOPS_DWELL,
+        dwellTime: DEFAULT_ADD_STOPS_DWELL,
         fromStop: null,
         routes: null,
         segments: [],
         feed: feedId,
-        speed: DEFAULT_ADD_STOPS_SPEED,
+        segmentSpeeds: [],
         toStop: null
       }
     case ADD_TRIP_PATTERN:
